updated login/register v2
