+++ src/main/java/com/dkt/cgshoppii/controller/LoginServlet.java
@@ -1,7 +1,8 @@
 package com.dkt.cgshoppii.controller;
 
+import com.dkt.cgshoppii.model.entity.User;
 import com.dkt.cgshoppii.service.IUserService;
-import com.dkt.cgshoppii.service.UserService;
+import com.dkt.cgshoppii.service.impl.UserService;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -9,6 +10,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.List;
 
 @WebServlet(name = "loginServlet", value = "/login")
 public class LoginServlet extends HttpServlet {
@@ -24,17 +26,20 @@ public class LoginServlet extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         String username = request.getParameter("username");
         String password = request.getParameter("password");
-        String message;
+        List<User> users = userService.findAllUsers();
+        String message = "Invalid credentials!"; // Default message for invalid credentials
 
-        // Validate login credentials (simple example)
-        if (username.equals("admin") && password.equals("admin")) {
-            message = "Login successful!";
-            request.getRequestDispatcher("WEB-INF/view/product/productList.jsp").forward(request, response);
-        } else {
-            message = "Invalid credentials!";
-            request.getRequestDispatcher("WEB-INF/view/user/Login.jsp").forward(request, response);
+        // Validate login credentials
+        for (User user : users) {
+            if (username.equals(user.getUsername()) && password.equals(user.getPassword())) {
+                message = "Login successful!";
+                request.setAttribute("message", message); // Set the success message
+                request.getRequestDispatcher("WEB-INF/view/product/productList.jsp").forward(request, response);
+                return; // Exit after successful login
+            }
         }
 
+        // If no user matched, set the error message
         request.setAttribute("message", message);
         request.getRequestDispatcher("WEB-INF/view/user/Login.jsp").forward(request, response);
     }
