--- src/main/java/com/dkt/cgshoppii/dao/UserDAO.java
@@ -1,78 +1,0 @@
-package com.dkt.cgshoppii.dao;
-
-import com.dkt.cgshoppii.connection.JDBCConnection;
-import com.dkt.cgshoppii.model.entity.User;
-
-import java.sql.*;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-
-public class UserDAO implements IUserDAO {
-    @Override
-    public List<User> findAllUsers() {
-        List<User> users = new ArrayList<>();
-        try (Connection connection = JDBCConnection.getConnection()) {
-            PreparedStatement ps = connection.prepareStatement("SELECT * FROM users");
-            ResultSet resultSet = ps.executeQuery();
-            while (resultSet.next()) {
-                int id = resultSet.getInt("id");
-                String username = resultSet.getString("username");
-                String password = resultSet.getString("password");
-                String role = resultSet.getString("role");
-                String fullName = resultSet.getString("full_name");
-                String phone = resultSet.getString("phone");
-                String email = resultSet.getString("email");
-                String address = resultSet.getString("address");
-                User user = new User(id, username, password, role, fullName, phone, email, address, null, null);
-                users.add(user);
-            }
-        } catch (SQLException e) {
-            throw new RuntimeException("Error retrieving users: " + e.getMessage(), e);
-        }
-        return users;
-    }
-
-    @Override
-    public Optional<User> findUserByUsername(String username) {
-        try (Connection connection = JDBCConnection.getConnection()) {
-            PreparedStatement ps = connection.prepareStatement("SELECT * FROM users WHERE username = ?");
-            ps.setString(1, username);
-            ResultSet resultSet = ps.executeQuery();
-
-            if (resultSet.next()) {
-                int id = resultSet.getInt("id");
-                String password = resultSet.getString("password");
-                String role = resultSet.getString("role");
-                String fullName = resultSet.getString("full_name");
-                String phone = resultSet.getString("phone");
-                String email = resultSet.getString("email");
-                String address = resultSet.getString("address");
-                User user = new User(id, username, password, role, fullName, phone, email, address, null, null);
-                return Optional.of(user);
-            }
-        } catch (SQLException e) {
-            throw new RuntimeException("Error finding user by username: " + e.getMessage(), e);
-        }
-        return Optional.empty(); // Return empty if no user is found
-    }
-
-    @Override
-    public void saveUser(User user) {
-        try (Connection connection = JDBCConnection.getConnection()) {
-            String sql = "INSERT INTO users (username, password, role, full_name, phone, email, address) VALUES (?, ?, ?, ?, ?, ?, ?)";
-            PreparedStatement ps = connection.prepareStatement(sql);
-            ps.setString(1, user.getUsername());
-            ps.setString(2, user.getPassword()); // Make sure to hash the password in production
-            ps.setString(3, user.getRole()); // Save the role (User, Admin, None)
-            ps.setString(4, user.getFullName());
-            ps.setString(5, user.getPhone());
-            ps.setString(6, user.getEmail());
-            ps.setString(7, user.getAddress());
-            ps.executeUpdate();
-        } catch (SQLException e) {
-            e.printStackTrace();
-            throw new RuntimeException("Error saving user: " + e.getMessage(), e);
-        }
-    }
-}
